api: Add support for stale-while-revalidate and ETag/304
These translations very rarely update. The Last-Modified thing
was not very useful because we had no timestamp to base it on,
so to be safe it was set to 5min, which meant a re-download after
5 min.

By using ETag instead of Last-Modified for the 304 checks, it will
never redownload if the browser already has a cached copy that is
identical, no matter how old.

Also, try to keep the 304 checks out of the critical path for
gadgets loading, by using "stale-while-revalidate". This means
that after the 5min blind cache expires, the next page view,
if within the hour, will still re-use it blindly and then perform
the check (or update) in the background.